Move the test for _write_download_file() to test_magic_folder module
